fix imports and rename types
