--- src/components/MainPageTitle/MainPageTitle.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { MainPageTitleWrapper } from './MainPageTitle.styled';
-
-const MainPageTitle = () => {
-  return (
-    <MainPageTitleWrapper>
-      <h1>So Yummy</h1>
-      <p>Discover the most delicious recipes</p>
-    </MainPageTitleWrapper>
-  );
-};
-
-export default MainPageTitle;
