+++ src/pages/CategoriesPage/CategoriesPage.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect, Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import MainContainer from '../../components/MainContainer';
+import MainContainer from 'components/MainContainer';
 import CategoriesList from 'components/CategoriesByName/CategoriesList';
 import ReusableTitle from 'components/ReusableComponents/ReusableTitle';
 import Loader from 'components/Loader';
